Reduce compiler warnings (#2456)
* Exclude -fno-rtti option from compilation of c files

* Change string to char type for equal sign in android input

* Removed unused variables

* Remove double const declaration

* Move platform specific variable in same scope as usage

* Fix unitiallized/no return path warnings

* Fix type conversion warnings

* Use correct path errno.h include

* Explicitly ignore fread() return value

* Remove obsolete -Wno-warn-absolute-paths command line option for web build

* Fixed compiler warning for win32

* Explicitly use float versions of atan2 and assign

* More type conversion warning fixes in hashtable.h for win32

* Don’t define be64toh if it already exists (it does on Linux)

* Win32 no longer requires declaration of alloca()

* Moved dmGui::DEFAULT_LAYER and DEFAULT_LAYOUT value to cpp file

* Use uint32_t for active_attachment in rig

* Move dmGui::DEFAULT_LAYER inside of gui.cpp

* Review fixes

+++ engine/resource/src/resource.cpp
@@ -203,7 +203,7 @@ static void HttpHeader(dmHttpClient::HResponse response, void* user_data, int st
         if (factory->m_HttpContentLength < 0) {
             dmLogError("Content-Length negative (%d)", factory->m_HttpContentLength);
         } else {
-            if (factory->m_HttpBuffer->Capacity() < factory->m_HttpContentLength) {
+            if (factory->m_HttpBuffer->Capacity() < (uint32_t)factory->m_HttpContentLength) {
                 factory->m_HttpBuffer->SetCapacity(factory->m_HttpContentLength);
             }
             factory->m_HttpBuffer->SetSize(0);
@@ -1050,7 +1050,7 @@ static Result DoLoadResourceLocked(HFactory factory, const char* path, const cha
             return factory->m_HttpFactoryResult;
 
         // Only check content-length if status != 304 (NOT MODIFIED)
-        if (factory->m_HttpStatus != 304 && factory->m_HttpContentLength != -1 && factory->m_HttpContentLength != factory->m_HttpTotalBytesStreamed)
+        if (factory->m_HttpStatus != 304 && factory->m_HttpContentLength != -1 && factory->m_HttpContentLength != (int32_t)factory->m_HttpTotalBytesStreamed)
         {
             dmLogError("Expected content length differs from actually streamed for resource %s (%d != %d)", factory_path, factory->m_HttpContentLength, factory->m_HttpTotalBytesStreamed);
         }
